🐛 fix: synchronized msw init
sc에서 안되는건 어쩔 수 없는데
cc에서 msw 로드되기 전에 fetch 할 경우 정상 작동 안됨 -> init 동기화
