+++ app/LoadMsw.ts
@@ -1,7 +1,8 @@
 'use client';
+
 export default function LoadMsw() {
-  if (process.env.NEXT_PUBLIC_API_MOCKING === 'enabled') {
-    require('@/mocks');
+  if (process.env.NEXT_PUBLIC_API_MOCKING === 'enabled' && typeof window === 'object') {
+    require('@/mocks/browserInit');
   }
   return null;
 }
