+++ instrumentation.ts
@@ -1,6 +1,6 @@
 export async function register() {
   if (process.env.NEXT_RUNTIME === 'nodejs' && process.env.NEXT_PUBLIC_API_MOCKING === 'enabled') {
-    require('mocks');
+    require('@/mocks/serverInit');
     console.log('[instrumentation][register] API mocking enabled, starting.');
   }
 }
