+++ mocks/browserInit.ts
@@ -1,0 +1,10 @@
+import { worker } from './browser';
+
+function initMocks() {
+  worker.start({ onUnhandledRequest: 'bypass' });
+  console.log('Mocks worker started');
+}
+
+initMocks();
+
+export {};
