+++ mocks/serverInit.ts
@@ -1,0 +1,10 @@
+import { server } from './server';
+
+async function initMocks() {
+  server.listen({ onUnhandledRequest: 'bypass' });
+  console.log('Mocks server started');
+}
+
+initMocks();
+
+export {};
